🌀 #10 Phase 2 : 200줄이 넘는 코드량의 파일들에 대한  `코드 분리` 리팩터링
