+++ frontend/src/widgets/event-manager/index.tsx
@@ -3,16 +3,22 @@ import { useEventStore } from '@/entities/event/model';
 import { Event } from '@/entities';
 import { EventForm } from '@/features/event-form';
 import { ParticipantManager } from '@/features/participant-manager';
-import { Edit, Trash2, Plus, Users, Calendar, MapPin } from 'lucide-react';
-import { Modal, message, Popconfirm, Badge } from 'antd';
+import { Modal, message } from 'antd';
 import { EmptyState } from '@/shared/ui/Empty';
 import { LoadingSpinner } from '@/shared/ui/Spinner';
-import { DefaultButton } from '@/shared/ui/Button';
+import { EventManagerHeader } from '@/widgets/event-manager/ui/EventManagerHeader';
+import { EventCard } from '@/widgets/event-manager/ui/EventCard';
 
 interface EventManagerProps {
   organizationId: string;
 }
 
+/**
+ * 이벤트 관리 위젯
+ * - 이벤트 목록 표시
+ * - 이벤트 추가/수정/삭제
+ * - 참가자 관리
+ */
 export function EventManager({ organizationId }: EventManagerProps) {
   const { events, loading, deleteEvent } = useEventStore();
   const [isModalVisible, setIsModalVisible] = useState(false);
@@ -80,38 +86,15 @@ export function EventManager({ organizationId }: EventManagerProps) {
 
   return (
     <div className="space-y-6">
-      <div className="flex justify-between items-center">
-        <div className="flex items-center gap-3">
-          <h2 className="text-2xl font-bold text-gray-900">모임 목록</h2>
-          <div className="px-3 py-1 text-sm font-medium flex items-center gap-1">
-            <span>총</span>
-            <Badge
-              count={organizationEvents.length}
-              style={{ backgroundColor: '#667eea' }}
-              dot={false}
-            />
-            개
-          </div>
-        </div>
-        <DefaultButton
-          onClick={handleAddEvent}
-          icon={<Plus size={16} />}
-          type="primary"
-        >
-          새 모임 추가
-        </DefaultButton>
-      </div>
+      {/* 헤더 */}
+      <EventManagerHeader
+        totalEvents={organizationEvents.length}
+        searchTerm={searchTerm}
+        onSearchChange={setSearchTerm}
+        onAddEvent={handleAddEvent}
+      />
 
-      <div className="flex gap-4">
-        <input
-          type="text"
-          placeholder="모임명으로 검색..."
-          value={searchTerm}
-          onChange={(e) => setSearchTerm(e.target.value)}
-          className="flex-1 px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-primary"
-        />
-      </div>
-
+      {/* 이벤트 목록 */}
       {filteredEvents.length === 0 ? (
         <EmptyState
           title="등록된 모임이 없습니다"
@@ -120,66 +103,13 @@ export function EventManager({ organizationId }: EventManagerProps) {
       ) : (
         <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
           {filteredEvents.map((event) => (
-            <div
+            <EventCard
               key={event._id}
-              className="bg-white rounded-lg shadow-md p-6 border border-gray-200 hover:shadow-lg transition-shadow"
-            >
-              <div className="flex justify-between items-start mb-4">
-                <h3 className="text-lg font-semibold text-gray-900">
-                  {event.title}
-                </h3>
-                <div className="flex gap-2">
-                  <button
-                    onClick={() => handleEditEvent(event)}
-                    className="p-2 text-gray-600 hover:text-primary-500 hover:bg-gray-100 rounded-lg transition-colors"
-                    title="모임 편집"
-                  >
-                    <Edit size={16} />
-                  </button>
-                  <Popconfirm
-                    title="정말로 이 모임을 삭제하시겠습니까?"
-                    onConfirm={() => handleDeleteEvent(event)}
-                    okText="삭제"
-                    cancelText="취소"
-                  >
-                    <button
-                      className="p-2 text-gray-600 hover:text-red-600 hover:bg-red-50 rounded-lg transition-colors"
-                      title="모임 삭제"
-                    >
-                      <Trash2 size={16} />
-                    </button>
-                  </Popconfirm>
-                </div>
-              </div>
-
-              <p className="text-gray-600 mb-4 line-clamp-2">
-                {event.description}
-              </p>
-
-              <div className="space-y-2 text-sm text-gray-500">
-                <div className="flex items-center gap-2">
-                  <Calendar size={14} />
-                  <span>{new Date(event.date).toLocaleDateString()}</span>
-                </div>
-                <div className="flex items-center gap-2">
-                  <MapPin size={14} />
-                  <span>{event.location}</span>
-                </div>
-                <div className="flex items-center gap-2">
-                  <Users size={14} />
-                  <span>{event.currentParticipants}명 참여</span>
-                </div>
-              </div>
-
-              <div className="mt-4 pt-4 border-t border-gray-100">
-                <button
-                  onClick={() => handleManageParticipants(event)}
-                  className="w-full px-4 py-2 bg-gray-100 text-gray-700 rounded-lg hover:bg-gray-200 transition-colors"
-                >
-                  참가자 관리
-                </button>
-              </div>
-            </div>
+              event={event}
+              onEdit={handleEditEvent}
+              onDelete={handleDeleteEvent}
+              onManageParticipants={handleManageParticipants}
+            />
           ))}
         </div>
       )}
